test(extension): e2e - updated after review
