--- packages/e2e-tests/src/assert/nftAddNftToFolderAssert.ts
@@ -1,17 +1,0 @@
-import nftAddNftToFolderPage from '../elements/NFTs/nftAddNftToFolderPage';
-
-class NftAddNftToFolderAssert {
-  async assertSeeAddNftButton() {
-    await nftAddNftToFolderPage.placeholderItem.waitForDisplayed();
-    await nftAddNftToFolderPage.placeholderItem.waitForClickable();
-  }
-
-  async assertSeeNftAddNftToFolderPage() {
-    await nftAddNftToFolderPage.drawerContent.waitForDisplayed();
-    await nftAddNftToFolderPage.drawerHeaderTitle.waitForDisplayed();
-    await nftAddNftToFolderPage.selectedFolderTitle.waitForDisplayed();
-    await nftAddNftToFolderPage.selectedFolderNftCount.waitForDisplayed();
-  }
-}
-
-export default new NftAddNftToFolderAssert();
